pulling compressor out.
